Move oauth from pkg to model
