+++ web/apps/apps.go
@@ -12,10 +12,9 @@ import (
 	"path"
 	"strconv"
 
-	"github.com/cozy/cozy-stack/pkg/apps/appfs"
-	"github.com/cozy/cozy-stack/pkg/oauth"
-
+	"github.com/cozy/cozy-stack/model/oauth"
 	"github.com/cozy/cozy-stack/pkg/apps"
+	"github.com/cozy/cozy-stack/pkg/apps/appfs"
 	"github.com/cozy/cozy-stack/pkg/consts"
 	"github.com/cozy/cozy-stack/pkg/instance"
 	"github.com/cozy/cozy-stack/pkg/jsonapi"
