+++ worker/push/push.go
@@ -13,9 +13,9 @@ import (
 	"time"
 
 	"github.com/cozy/cozy-stack/model/notification/center"
+	"github.com/cozy/cozy-stack/model/oauth"
 	"github.com/cozy/cozy-stack/pkg/config/config"
 	"github.com/cozy/cozy-stack/pkg/jobs"
-	"github.com/cozy/cozy-stack/pkg/oauth"
 	"github.com/sirupsen/logrus"
 
 	fcm "github.com/appleboy/go-fcm"
